+++ tests/test_cloudpath_file_io.py
@@ -198,9 +198,13 @@ def test_glob(glob_test_dirs):
     #  https://github.com/python/cpython/blob/7ffe7ba30fc051014977c6f393c51e57e71a6648/Lib/test/test_pathlib.py#L1634-L1720
 
     def _check_glob(pattern, glob_method, **kwargs):
+        local_pattern = kwargs.pop("local_pattern", None)
+
         _assert_glob_results_match(
             getattr(cloud_root, glob_method)(pattern, **kwargs),
-            getattr(local_root, glob_method)(pattern, **kwargs),
+            getattr(local_root, glob_method)(
+                pattern if local_pattern is None else local_pattern, **kwargs
+            ),
             cloud_root,
             local_root,
         )
@@ -214,6 +218,10 @@ def test_glob(glob_test_dirs):
     _check_glob("*A", "glob")
     _check_glob("*B/*", "glob")
     _check_glob("*/fileB", "glob")
+    _check_glob(cloud_root / "**/*", "glob", local_pattern="**/*")
+
+    if sys.version_info >= (3, 13):
+        _check_glob(PurePosixPath("**/*"), "glob")
 
     # rglob_common
     _check_glob("*", "rglob")
@@ -222,6 +230,10 @@ def test_glob(glob_test_dirs):
     _check_glob("*/fileA", "rglob")
     _check_glob("*/fileB", "rglob")
     _check_glob("file*", "rglob")
+    _check_glob(cloud_root / "*", "rglob", local_pattern="*")
+
+    if sys.version_info >= (3, 13):
+        _check_glob(PurePosixPath("*"), "rglob")
 
     dir_c_cloud = cloud_root / "dirC"
     dir_c_local = local_root / "dirC"